[FIX] sale_stock_renting: add reference to manual rental transfer
Steps to reproduce:

- Create a rentable, storable product
- Create and confirm a rental order for 1 unit of this product
- Click on pickup
- Inventory > Reporting > Moves History
> Your pickup move appears without any reference

Cause of the issue:

The issue has been introduced in d0c1e7845feeee1c2e85a21b5d40570d051458d3
which purpose was to remove the `name` field of the `stock.move` model.
However, the `_compute_reference` compute method use to rely on this
`move.name` to propagate the info that the move was related to a rental
order (since there is no picking). Indeed prior to saas-18.4, the
compute method was:
https://github.com/odoo/odoo/blob/404cb10283cbc706eae67dd793ced363273f3602/addons/stock/models/stock_move.py#L325-L328
And the reference to the rental order was set on the move at pickup:
https://github.com/odoo/enterprise/blob/1466a0139ee64ecd059738bc414f6e7e5a9f4354/sale_stock_renting/models/sale_order_line.py#L303-L313
https://github.com/odoo/enterprise/blob/1466a0139ee64ecd059738bc414f6e7e5a9f4354/sale_stock_renting/models/sale_order_line.py#L248-L257

Fix:

Since the reference field is a computed and stored fields and since some
of its dependencies are set at creation the rental move we can not set
the `reference` directly in the creation of the record as we used to do
for its name since the compute method will then override and erase
or reference.

opw-5385004

closes odoo/enterprise#101810

X-original-commit: faa4dbd8a8b066afc1d4c999d1301daef2bef0fa
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>