[FIX] l10n_au_hr_payroll_account: prevent error when creating payslip draft entries
Currently, an error occurs when generating draft entries for the Australian
payslips when the employee is configured to report values in BAS–W3.

Steps to reproduce:
- Install the `l10n_au_hr_payroll_account` module and switch
  to `My Australian Company`.
- Create an employee with all necessary payroll information and
  enable `Report in BAS – W3`.
- Go to Payroll Settings and complete the `Payroll Onboarding`.
- Navigate to Payroll > Payslips > Payslips and open any draft payslip.
- Click Create Draft Entry.

(Refer [1] for steps to produce.)

Error:
TypeError: 'int' object is not iterable

Root Cause:
At [2], the code tries to append an integer(`w3_tag.id`) to a list(`tag_ids`)
using `+=`, which causes the error.

Fix:
This commit prevents an error when creating a draft entry by wrapping the
integer in a list before adding it to `tag_ids`, similar to [3].

[1]:
https://drive.google.com/file/d/1XGUUYCpmB3ocwXSWkxgFV7vjyAvRsHS6/view

[2]:
https://github.com/odoo/enterprise/blob/d9c864a5a55dd45f6fed794a809a5aefe26e7699/l10n_au_hr_payroll_account/models/hr_payslip_line.py#L23

[3]:
https://github.com/odoo/enterprise/blob/d9c864a5a55dd45f6fed794a809a5aefe26e7699/l10n_au_hr_payroll_account/models/hr_payslip_line.py#L27

opw-5370293

closes odoo/enterprise#101811

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>