[FIX] sign: prevent redirect to signing mode when sender is also a signer
When creating a Sign Request where the sender is also one of the signers, clicking Send in the wizard incorrectly redirected to the document preview/signing interface.
This fix ensures that clicking Send always closes the wizard and does not trigger any redirect, regardless of the sender’s signer status.

task-5408155

closes odoo/enterprise#101816

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>