[FIX] l10n_se_bban: fix iso20022 xml
Here https://github.com/odoo/enterprise/pull/95463, we add the possibility to export
batch payments with mixed IBAN and Bankgiro/Plusgiro/BBAN payments, but this introduced
few bug in the xml format.

The reason is, we were using new custom logics and not the main one. The problem is
the custom logics is wrong, not the main one.

This commit remove most of the custom logics we added and use all the main one.

This has been done by:
    1 - Reverting the original commit
    2 - Adding only the zip file generation, passing a context key to know if we are
        with bban or iban payments.

opw-5181340

closes odoo/enterprise#102050

X-original-commit: e0cf114e0496d830842d974b8e5a5dbc380007ba
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Igor Bertrand (igbe) <igbe@odoo.com>