[FIX] hr_appraisal: don't reset feedback upon appraisal reopen
Steps to reproduce:
1- Create new appraisal for Abigail Peterson
2- Confirm appraisal
3- Edit employee and manager feedbacks
4- Set the final rating for the appraisal
5- Mark appraisal as done
6- Reopen appraisal

Issue:
The feedbacks of both the employee and manager are reset to the default
template. They should be unchanged.

Cause:
Both the reopen and reset buttons used to call the the same action "action_back"
which would set the state of the appraisal to "1_new", there is a compute
that is triggered by the change in state which updates the appraisal's
department_id which in turn triggers two other computes for the employee
and manager feedbacks, those computes have a condition that if the state
of the appraisal is '1_new', then the feedbacks should be reset to the
templates.

Fix:
Make each button call its own action and make the reopen action set the
state to '2_pending' instead of '1_new'. Furthermore make the reset action
(which is called action_back) also reset the final rating for the appraisal.

Task-5411661

closes odoo/enterprise#102021

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>