[FIX] iot: traceback removing IoT records
For an unknown reason, when removing an IoT record,
Odoo was evaluating only partly an `invisible` tag.

The condition was `not version or version[0] == 'W'`.
When deleting, `not version` was still false, but
`version` in `version[0]` was undefined, resulting
in a traceback.

This is fixed using `'W' in version` instead.

opw-5352092

closes odoo/enterprise#102065

Signed-off-by: Max Whale (mawh) <mawh@odoo.com>