[FIX] iot: disable webrtc after failure
Before this commit, WebRTC was always used as the first protocol when
communicating with the IoT box. However, in an environment where it was
not working, it would take 6 seconds to fallback to longpolling on every
request.

After this commit, if WebRTC fails, we don't try to use it again for 20
minutes.

task-5412223

closes odoo/enterprise#102110

Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>