[FIX] pos_blackbox_be: undefined employee unlocking pos
When a PoS was configured with "login with employees" and
a blackbox, locking the pos then trying to select an employee
raised a traceback.

The PoS was trying to update the clock in date, but the employee
was undefined.

This is now fixed checking whether if the employee is set.

closes odoo/enterprise#102137

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>