[FIX] iot: idempotency fix
Currently multiple clients report double order receipts printing. This PR fixes the issue where due to slow network connection a request would be sent to the iot box but the iot box didn't reply in time to confirm the action finish.

The user would then get an error showing 'printing failed' (due to a timout). If he cliks on retry the iot box would still print the previous receipt and then receive the new "retry" request with now a new action uuid which would also be printed because uuid is different from the 1st request.

This PR adds a consistent uuid for both the initial and all the subsequent retry requests so that double actions would never be done by the iot box.

closes odoo/enterprise#102266

Related: odoo/odoo#240253
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>