[FIX] account_reports: prevent error when sending the customer statement
Currently, an error occurs when a user sends a customer statement.

Steps to Reproduce:
 - Install the account_reports module.
 - Go to Accounting > Customers > Customers.
 - Switch to List view, select a customer, then click on Actions >
   Open Customer Statements.
 - Click Send, remove the email template, add a subject, and then
   click Print & Send.

ValueError: Expected singleton: mail.template()

After [this commit], which checks whether the template has an email_from,
when a user sends the customer statement and removes the email template,
it still tries to access the template to fetch email_from for a particular
customer [1]. If no email template is used, this results in an error [2] when
going to extract the email_from.

This commit ensures that email_from is taken from the email template
if one is used; otherwise, it uses the current user's email address [3],
which matches the default behavior.

[this commit]: https://github.com/odoo/enterprise/commit/22c46e4f63e7b2dc0eee16fc807656cd4de21fb7

[1]: https://github.com/odoo/enterprise/blob/dc4d5633407ccc728d30de5ce2072a80c16b2766/account_reports/wizard/account_report_send.py#L246

[2]: https://github.com/odoo/odoo/blob/0dabb221225fba96c0e55779afadd9f12b369777/addons/mail/models/mail_render_mixin.py#L691-L693

[3]: https://github.com/odoo/odoo/blob/0dabb221225fba96c0e55779afadd9f12b369777/addons/mail/models/mail_thread.py#L2891-L2892

sentry-7106576491

closes odoo/enterprise#102682

X-original-commit: 3df585779e6ede664279331c4531043688dadf17
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Ashutosh Sharma (assh) <assh@odoo.com>