[FIX] l10n_it_pos: use price unit instead of line total
Currently, when buying multiple quantities of the same product, the
unit price value sent to the italian printer is incorrect.

Steps to reproduce:
-------------------
* Set up italian printer for one shop
* Open shop
* Add a product to the order with a qty 3 and a price unit of 1
* Pay the order
* Print italian receipt
> The price total for the 3 product says 9 instead of 3

Why the fix:
------------
When computing the order lines details we were using `total_included`
for the unit price which takes into account the quantity. Per definiton,
it's not the price for 1 unit.

opw-5404877

closes odoo/enterprise#102457

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>