[FIX] account_reports: don't over-indent custom groupby sublines with complex domains
When a custom groupby defined a domain with more than one element in it (typically, conditions on multiple fields, possibly with '|' or '&' operations), if groupby lines needed to be generated for subsequent groupby, the generated groupby lines had a too high level, and ended up being over-indented.

This was because the groupby expansion computed the number of higher-level groupby lines as the length of the domain needed to restrict their content. This was too naïve, and won't work well with custom groupbys, which can define longer (or even empty) domains.

This was spotted in master (19.1), with the refactored EC Sales list, which now uses a custom engine and a specific custom groupby map.

closes odoo/enterprise#102240

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>