[FIX] account_accountant: deferred misc entry
The commit 42f823d6b8aa3d1cd171ae1603549ee95fc9d0f0 allows to use
deferred on misc entries. However, there are many places in the code
that were not updated. Usually they were in the form of `if move_type is
 sale, then deferred_type = income, else expense`.

However we cannot rely on the move_type anymore for misc entries,
because it will always take the `else` branch of the condition. Instead,
 if we have a misc entry, we should rely on the account type of the line
  that is being deferred, so we have more granularity.

For this, we now compute the deferral account/journal for each line, and
 not per move. The logic inside the computation remains the same.

Steps to reproduce:
1. Create a misc entry with two deferred lines (one expense, one revenue)
2. Post it
3. Check the generated deferrals, they all use the same deferred account
 and journal even though we have different account types

opw-5194305

closes odoo/enterprise#102572

X-original-commit: 3e6d2f3ca7e2d4e940f2c2022f816202c72cbd1b
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>