[FIX] web_studio: Handle insignificant changes in `save_report`
Example of steps:
- Install sale_management and web_studio
- Try to edit PDF Quote report
- add /4 columns somewhere above the table
- try to save
- Error: Document is empty

In this case, for some reason `html_editor` edits the external_report by
changing the order of attributes on a node, which causes a diff, but not
relevant for studio and won't produce any xpath operations.

This commit handles this case by initializing the Studio view to
`<data/>`.

opw-5351588

closes odoo/enterprise#102416

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>