[FIX] ai: traceback when AI action prompt is empty
Prior to this commit, when the AI action prompt was left empty, we would
get a traceback when that AI action was executed. The traceback would
occurs because the prompt is an HTML field with the field selector plugin,
which means that it requires parsing before it's sent to the LLM. There
was no check during parsing whether the field is empty or not and thus a
traceback would occur during processing.

In this commit we added a check for whether the prompt field is empty or
not and if it is, we skip parsing altogether. We still call the LLM but
without the final-prompt. The pre-prompt with contextual information is
still sent thus the LLM can still perform the action if it's simple enough.

Also, if the AI server action prompt is ambiguous and the LLM "feels"
like it needs additional input to complete its task, it might respond
with an additional query to the user instead of calling one of its available tools.

Thus, in this commit, we add an explicit instruction to the AI Server action
pre-prompt which should convey to the LLM that it will not be able to get
any additional input from the user and it should assume that any important
information for performing its task will be hardcoded in its available tools.

Task-5379758

closes odoo/enterprise#102441

Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>