[FIX] l10n_be_reports_client_nihil: don't set client_nihil to True for non-end-of-year-periods
When the client_nihil field was hidden from the wizard, its computation always set it to True. This was wrong : this field can only be True in the XML when in the last period of the year, so when should_display_client_nihil is True as well. In every other cases, it needs to be False.

opw-5184056

closes odoo/enterprise#102454

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>