[FIX] l10n_cl_edi: fix ensure_one on _get_start_number
In #92208 the CAF system was improved to find the next starting value in
the right sequence if there are multiple in the system. If no CAFs are
found, it will reset to the document types starting number.

This worked except in the case where there are multiple CAFs that are
currently marked active in a document type. As it tries to access
`caf.start_nb` it hits an ensure_one() which throws a traceback whenever
you open most account.moves on the DB.

Steps to reproduce:
- Modify the CAF for Doc Type 33 (Electronic Invoices) to have a smaller
  range than 1 - 999,999 (1 - 5)
- Create Two new CAF files also for Doc Type 33 that start after this
  value (6 - 10 and 11 - 15 for example).
- Mark the original to be spent via cancelling it and try to open an
  invoice. It will find both of the new CAFs and try to get the start_nb
  of the recordset.

In discussion with the PO, when we have multiple CAFs, we should pick by
the lowest start number as it will allow for any gaps that might exist
be filled.

opw-5414350

closes odoo/enterprise#102486

X-original-commit: b49472b673ee22325c88f9ee80c4e816c9f128ad
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: Andrew Gavgavian (andg) <andg@odoo.com>