[FIX] whatsapp: match position of 'view-contact' with 'livechat-info'
The thread action 'view-contact' is somewhat like 'livechat-info',
in that this action redirects to relevant info to the whatsapp user
and livechat-info is a panel with all info on the visitor and more.

The position of both buttons were the total opposite in discuss app:
- 'livechat-info' was put as the 1st panel one before panels
- 'view-contact' was dead last.

There are differences, with 'livechat-info' being a panel and
'view-contact' a redirect button. However, each are important buttons
for the type of conversation that justify them being in their own
group right now. Since the buttons are mutually exclusive by the
channel type, there's reason to have them share the same position.

This commit moves the position of 'view-contact' in discuss header
thread actions to be the same as 'livechat-info', that is as the 1st
action in the list. This makes its position consistent with livechat
info, thus agents of live chat and whatsapp would appreciate the 2
button placement being consistent.

closes odoo/enterprise#102479

Signed-off-by: Didier Debondt (did) <did@odoo.com>