[IMP] account_reports,l10n_mx_reports_closing: no initial balance for equity_unaffected
Following the changes introduced in odoo#240742 and in enterprise#95062,
this commit adapts the l10n_mx_reports_closing tests, as the equity_unaffected
account amounts change if include_initial_balance = False.

This commit also moves the domain logic for the Undistributed Line to a dedicated
function, making it simpler to customize this behavior if needed.

no-task

closes odoo/enterprise#102549

Related: odoo/odoo#240742
Signed-off-by: William André (wan) <wan@odoo.com>