[ADD] l10n_pe_reports_lib: Peruvian 5th and 6th Inventory and Balance reports
Adding 'Inventory and Balance' export option to the Peruvian general ledger
report, and fields required for these reports to account_account and
res_company. Report lines are handled with direct SQL queries since no UI
display is required for this one, and some elements needed are outside of the
scope of the usual reports model.
task-4057152

closes odoo/enterprise#102356

X-original-commit: 56d1cfb1e36969cce1d9c42c96088c367d9cdd06
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>