[FIX] pos_*: availability of preset in self
*: pos_self_order, pos_online_payment_self_order

Two things will change with this PR.

First, before this commit, when a pos config had only one available preset (and set as default) and this preset was not available in self, this preset was used as the default one.
But now, it's not the case anymore. We only use it if, and only if, it is available in self.

The second thing was to fix the "Dine In" preset availability in the eating location choices. A dine-in preset is a preset which have service_at == table as attribute.

If this preset (is `available_in_self`) and has `service_at === table`, we only show it when there is a table_identifier in the URL. So, it'll never be shown in any self without table_identifier.

If user wants to see this preset in the kiosk or in the mobile menu without table_identifier, user needs to change the `service_at` field at `counter` (or `delivery`)  and put the product as `available in self`

closes odoo/enterprise#103063

Task-id: 5190541
Related: odoo/odoo#234522
Signed-off-by: David Monnom (moda) <moda@odoo.com>